fix: moved all regression tests into its own folder
