+++ cypress/integration/Regression/EditAndBranching/VerifyEditTrainingControlsAndLabels.spec.js
