+++ cypress/integration/Regression/EndSession/EndSession.spec.js
@@ -3,12 +3,12 @@
  * Licensed under the MIT License.
 */
 
-import * as homePage from '../../support/components/HomePage'
-import * as models from '../../support/Models'
-import * as modelPage from '../../support/components/ModelPage'
-import * as train from '../../support/Train'
-import * as scorerModal from '../../support/components/ScorerModal'
-import * as helpers from '../../support/Helpers'
+import * as homePage from '../../../support/components/HomePage'
+import * as models from '../../../support/Models'
+import * as modelPage from '../../../support/components/ModelPage'
+import * as train from '../../../support/Train'
+import * as scorerModal from '../../../support/components/ScorerModal'
+import * as helpers from '../../../support/Helpers'
 
 describe('End Session', () => {
   afterEach(helpers.SkipRemainingTestsOfSuiteIfFailed)
