+++ cypress/integration/Regression/ErrorHandling/ActionUnavailable.spec.js
@@ -3,12 +3,12 @@
  * Licensed under the MIT License.
 */
 
-import * as models from '../../support/Models'
-import * as modelPage from '../../support/components/ModelPage'
-import * as train from '../../support/Train'
-import * as trainDialogsGrid from '../../support/components/TrainDialogsGrid'
-import * as common from '../../support/Common'
-import * as helpers from '../../support/Helpers'
+import * as models from '../../../support/Models'
+import * as modelPage from '../../../support/components/ModelPage'
+import * as train from '../../../support/Train'
+import * as trainDialogsGrid from '../../../support/components/TrainDialogsGrid'
+import * as common from '../../../support/Common'
+import * as helpers from '../../../support/Helpers'
 
 describe('Action Unavailable - ErrorHandling', () => {
   afterEach(helpers.SkipRemainingTestsOfSuiteIfFailed)
