+++ cypress/integration/Regression/ErrorHandling/TwoConsecutiveUserInputs.spec.js
