+++ cypress/integration/Regression/ErrorHandling/WaitNonWait.spec.js
